[GH-32] feat(cache): add redis cache to product api

+++ camila-orchestrator/dev/compose/stop.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+# Example of usage: ./stop.sh removeImages=true
+
 set -o errexit # Exit on error. Append "|| true" if you expect an error.
 set -o errtrace # Exit on error inside any functions or subshells.
 set -o nounset # Do not allow use of undefined vars. Use ${VAR:-} to use an undefined VAR
@@ -9,15 +11,24 @@ SEPARATOR="\n ################################################## \n"
 
 cd "$(dirname "$0")"
 
+parameter="${1:-"removeImages=false"}"
+eval "${parameter}"
+echo "removeImages: ${removeImages:-}"
+
 __stopServices() {
   echo -e "${SEPARATOR} 🔨 Stop services ${SEPARATOR}"
-  docker-compose --file docker-compose.yml down --remove-orphans --volumes
+  docker-compose --file docker-compose.yml down --remove-orphans --volumes || true
 }
 
 __cleanEnvironment() {
   echo -e "${SEPARATOR} 🧹 Clean up ${SEPARATOR}"
   docker volume prune --force --all
-  docker images --filter reference='camila-*' --format '{{.Repository}}:{{.Tag}}' | xargs -I {} docker rmi -f {}
+
+  if [[ "${removeImages:-}" == "true" ]]; then
+    docker images --filter reference='camila-*' --format '{{.Repository}}:{{.Tag}}' | xargs -I {} docker rmi -f {}
+  else
+    echo -e "🚧 Skip remove images"
+  fi
 }
 
 main() {
